+++ Changelog.md
@@ -2,6 +2,25 @@
 
 ### Gazebo Physics 6.x.x (202X-XX-XX)
 
+### Gazebo Physics 6.1.1 (2022-11-07)
+
+1. Suppress MSVC warning that is not applicable
+    * [Pull request #446](https://github.com/gazebosim/gz-physics/pull/446)
+
+1. Fix construction of nested models 
+    * [Pull request #445](https://github.com/gazebosim/gz-physics/pull/445)
+
+### Gazebo Physics 6.1.0 (2022-10-27)
+
+1. Removed GetCollision API it's redundant
+    * [Pull request #444](https://github.com/gazebosim/gz-physics/pull/444)
+
+1. Add bullet-featherstone plugin
+    * [Pull request #373](https://github.com/gazebosim/gz-physics/pull/373)
+
+1. Fix bullet cmake, probably bad merge
+    * [Pull request #436](https://github.com/gazebosim/gz-physics/pull/436)
+
 ### Gazebo Physics 6.0.0 (2022-09-21)
 
 1. Improved documentation and instructions